// web/src/lib/api/generated/models/internal-products-product-option-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface InternalProductsProductOptionResponse {
    'additional_price'?: number;
    'id'?: string;
    'image_url'?: string;
    'name'?: string;
}